Create .sln file
I had resisted creating one due to the likelihood that I would forget to
change the project GUIDs when using the template, thereby making them
not-globally-unique.  However, the risks appear to be small (are there
any?) and the benefits of being able to use Dependabot and easily build
with tools expecting a sln (e.g. CI systems) are great.  Add one.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>