Mark UnitTests as CLSCompliant(false)
The value of CLS-compliance is minimal for the UnitTests assembly, which
is not a library, not likely to be used by any code, much less code in a
language other than C#.  The hassle of annotating individual classes
with CLSCompliant is not justified.  Suppressing CS3016 and pretending
the classes are CLSCompliant is also not justified.  Mark them as not
CLS-compliant with the CLSCompliant(false) attribute.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>