Revert "Bump TargetFramework to netstandard2.1"
Since .NET Framework 4.6.1 and above support netstandard2.0, and none
support netstandard2.1, it makes netstandard2.0 a desirable target
despite the nullability problems previously discussed.  It is
particularly useful for this template to support testing with .NET
Framework, which can easily be removed, but not as easily added (which
requires figuring out how to do it).

This reverts commit 41bc27dcabda04d84517d43266f65b78aa03cbe5.