Test with netcoreapp2.0
Applications targeting .NET Standard 2.0 may want to test with the
earliest supporting .NET Core version, which is .NET Core 2.0.  Target
this framework for testing.

Suppress NETSDK1138 using CheckEolTargetFramework.  Targeting
unsupported frameworks may be desirable for tests for many reasons.
The risks of unpatched security issues are also smaller in tests, which
are less likely to be exposed to potentially malicious actors.

Exclude the xunit.runner.visualstudio dependency, which is not supported
on netcore2.0.  Shouldn't be an issue, since the CI tests are not run
from the Visual Studio runner.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>