.github: run jobs simultaneously
Running a single test-primary job first was intended to avoid wasting
resources in the (common) case that failures are "all or none" across
matrix configurations.  However, since [GitHub Actions does not support
YAML anchors], this is painful and error-prone.

Since the matrix only defines 2 jobs, the gain is not worth the pain.
Use a single job definition with matrix.

[GitHub Actions does not support YAML]: https://github.community/t/16128

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>