.github: Upload each coverage file
When a test project has multiple TargetFrameworks, a coverage file is
produced for each.  Upload all coverage files to coverage services.

Note: I would have preferred the flag name include the TargetFramework,
but could not find a simple way to accomplish that.  Add a comment for
future work and use numbers for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>