UnitTests: Target netcoreapp2.1
Although .NET Core 2.0 is the earliest version which implements .NET
Standard 2.0, it is currently unsupported and xunit.runner.visualstudio
is not available.  Without it, `dotnet test` fails with:

> Unable to find /.../NetCoreProject/NetCoreProject/NetCoreProject.UnitTests/bin/Release/netcoreapp2.0/testhost.dll.  Please publish your test project and retry.

Although there may be a way to work around the issue, it's not worth
currently worth the effort.  Target .NET Core 2.1 instead.

Fixes: a32a04455d8d ("Test with netcoreapp2.0")

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>